[fasedtests] Rename fased -> fasedtests
