+++ sim/src/main/scala/fasedtests/Generator.scala
@@ -1,6 +1,6 @@
 //See LICENSE for license details.
 
-package firesim.fased
+package firesim.fasedtests
 
 import chisel3.internal.firrtl.{Port}
